Implement GSS_C_CHANNEL_BOUND_FLAG
Define a new channel-bound GSS return flag.  In the krb5 mech, set the
return flag if matching channel bindings were sent.  If the acceptor
indicates understanding of the channel-bound flag, allow the initiator
to omit channel bindings even if the acceptor provides them.  If the
acceptor does not indicate understanding of the channel-bound flag,
error out if the acceptor provides channel bindings and the initiator
does not, even if the initiator does not send a standard checksum.

[ghudson@mit.edu: simplified code changes; fleshed out commit message]

[iboukris: partly cherry-picked from PR 685]