RBCD: KDC: Handle local and cross-realm RBCD requests
For local request, that is when the impersonator and the proxy
target are in the same realm, the handling is similar to traditional
constrained delegation although the evidence ticket is not required
to be forwardable.

Then there are mainly three types of cross realm RBCD reuqests.
- KDC in impersonator realm.
- KDC in target proxy realm.
- KDC in transit realm.

The first two, are constrained delegation requests, in the sense
that the constrained-delegation flag is set and they come with a
second ticket. The third one comes as a regular cross-realm
request (for a krbtgt service), and besides the information in
the PAC in the authorization data, there is no indication of it
being used for constrained delegation. So the third type is
handled mainly in KDB plugin sign_authdata() handler.

Signed-off-by: Isaac Boukris <iboukris@gmail.com>