RBCD: Ignore forwardable flag in constrained delegation
With the introduction of RBCD, the lack of forwardable
flag no longer means the ticket cannot be used for a
constrained delegation request.

Instead, we should check in the PAC if the user is sensitive,
and error out in that case rather than making a failed request.
But we don't always have access to the PAC and we currently do
not have the code to retreive this attribute from the PAC.

Signed-off-by: Isaac Boukris <iboukris@gmail.com>